+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggsurvfit (development version)
 
+* Improved error message when `ggcuminc()` is called with `survfit.coxphms` objects. (#222)
+
 * Fixed `add_risktable()` to respect x-axis breaks when x-axis is duplicated. (#221)
 
 * Fixed `add_pvalue()` function to work with all outcomes in `ggcuminc()`, not just the first outcome level. (#239, @samrickman)